Refactor InputType package
