--- app/src/main/java/com/do55anto5/moviestreaming/core/enums/InputType.kt
@@ -1,6 +1,0 @@
-package com.do55anto5.moviestreaming.core.enums
-
-enum class InputType {
-    EMAIL,
-    PASSWORD
-}
