+++ app/src/main/java/com/do55anto5/moviestreaming/core/enums/input/InputType.kt
@@ -1,0 +1,6 @@
+package com.do55anto5.moviestreaming.core.enums.input
+
+enum class InputType {
+    EMAIL,
+    PASSWORD
+}
